fix: resolve all CI errors for PR #959
- Fix frontend lint errors: export PdfTableCell type, remove unused imports
- Fix lint warnings: split context files into separate components/hooks
  * ThemeContext -> ThemeContext.ts, ThemeProvider.tsx, useTheme.ts
  * SharedClockContext -> SharedClockContext.ts, SharedClockProvider.tsx, useSharedClock.ts
- Fix build errors: correct exportToXLSX import, add generatePayrollReport
- Fix ExportModal.tsx: update onExport prop type to include 'pdf'
- Fix sanitize.ts: use dompurify instead of isomorphic-dompurify
- Create src/util/types.ts for shared type definitions

Resolves CI errors: lint, build, and TypeScript compilation issues
