+++ src/components/ui/ThemeToggle.tsx
@@ -1,4 +1,4 @@
-import { useTheme } from "../../context/ThemeContext";
+import { useTheme } from "../../context/useTheme";
 
 export const ThemeToggle = () => {
   const { theme, toggleTheme } = useTheme();
