+++ src/context/SharedClockContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import type { SharedClockContextValue } from "./useSharedClock";
+
+export const SharedClockContext = createContext<SharedClockContextValue | null>(
+  null,
+);
