+++ src/context/ThemeContext.ts
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+import type { ThemeContextValue } from "./useTheme";
+
+export const ThemeContext = createContext<ThemeContextValue | null>(null);
