+++ src/util/types.ts
@@ -1,0 +1,10 @@
+export type PdfTableCell =
+  | string
+  | {
+      content: string;
+      colSpan?: number;
+      styles?: {
+        halign?: "left" | "center" | "right" | "justify";
+        fontStyle?: "normal" | "bold" | "italic" | "bolditalic";
+      };
+    };
